requirements.txt
gradio
torch
torchvision
Pillow
gdown
numpy
scipy
cmake
onnxruntime-gpu
opencv-python-headless